fix: find open failed if no search string
Since the search string is used to find the line offset to open the file
at, if there is no search string given the file would not open, as the
attempt to split the resulting string at a space character would return
null.

Similiarly, if the search term contained a space character, the file
path it would try to open would not exist.

Changed so the search string and resulting path are split at the last
space instead of the first, and if no split is found, to use an empty
search string and default to line 0.