feat: augment selection with time string
Can now also parse the `--time` flag with the selection parser, that
will include the time information in the Day union structure. This has
no change for any functions that did not make use of time, and adds
optional methods for accessing the index of the entry in the journal
page that corresponds to the given time.