feat(edit): add --path-only command
This is mainly for editor support. Instead of opening the file / path
from an editor within `nkt`, instead just prints the path to the
terminal so that it can be wrapped into existing editors. An editor can
now make a system call to `nkt e $args --path-only` to get the path of a
user query.

fixme: This only works for notes reliably. Would probably need some kind
of file watching mechanism to make this work for journals or tasklists.