fix: use const ptr to time when converting to date
Since the date needs a reference to the timezone instead of the timezone
itself, by passing the time by value the timezone in the date may refer
to an invalid address.

Fixed for now by using a const pointer to the time instead, so that the
reference is valid outside of the function body.

Needs reworking at some point to avoid future related issues.