feat: make Root constructor heap allocate
One of the problems thats encroaching is that passing pointers around
when getting collections means that the pointers can become invalidated
if e.g. the collection cache is grown or shrunk. This commit anticipates
letting the collections hold onto the root pointer, which must be heap
allocated to avoid becoming invalid, and then access their information
by index (DOD) instead of by pointer.