feat: collections to hold pointer to Root
To avoid the Root cache and the collection instances going out of sync,
the collections now hold the Root pointer and an ID. Whenever
information about the collection is needed, we go back to the root and
do a lookup in the cache -- if the cache has in any way changed (or
moved in memory), this avoids potentially invalidating pointers.

Performance comparison with `poop` shows the performance is unchanged
for the majority of operations.