feat(select): add collection_name:selector
Can now specify the collection name by using the
`collection_name:selector` syntax, which will hint to the selection
resolver what the name of the selection should be. This is designed to
make interactive use of `nkt` more fluid, and therefore is exclusive
of the `--flag` based collection specification: only one can be used to
hint the name of the collection. Attempting to use both will raise an
`AmbiguousSelection` error.

Added tests for the argument parsing.

Added help string in `select` to document the new syntax.